test: refactor evaluteEditor
