+++ test/ui/package.json
@@ -8,6 +8,7 @@
   },
   "devDependencies": {
     "@testing-library/dom": "^10.4.1",
+    "@types/codemirror": "^5.60.17",
     "@vitest/browser-playwright": "workspace:*",
     "happy-dom": "latest",
     "vitest": "workspace:*"
